Fix data structure for firestore
