+++ src/rescuetime/dailyReportUtils.ts
@@ -1,18 +1,13 @@
 import { find, sum, values, propEq } from 'ramda'
 
-import { parseEfficiency, parseOverview, parseActivity } from './weeklyReportUtils'
+import { parseEfficiency, parseOverview, parseActivity } from '../utils/parseUtils'
 import { mergeTimeByCategory } from '../utils/misc'
 import { getScaledActivityScore } from '../webhook/messageUtils'
 
-import { Efficiency, Overview, Activity } from '../@types/models'
+import { DailyDataSummaryParams } from '../@types/models'
 
 export type DailySummary = ReturnType<typeof getDayilDataSummary>
 export type CompareSummary = ReturnType<typeof compareWithYesterday>
-type DailyDataSummaryParams = {
-  efficiency: Efficiency
-  overview: Overview
-  activity: Activity
-}
 
 const TODAY_INDEX = 0
 const YESTERDAY_INDEX = 1
@@ -43,11 +38,7 @@ export const compareWithYesterday = (data: [DailySummary, DailySummary]) => {
   }
 }
 
-export const getDayilDataSummary = ({
-  efficiency,
-  overview,
-  activity,
-}: DailyDataSummaryParams) => {
+export const getDayilDataSummary = ({ efficiency, overview, activity }: DailyDataSummaryParams) => {
   const parsedEfficiency = mergeTimeByCategory(parseEfficiency(efficiency))
   const parsedOverview = find(propEq('category', 'Software Development'), parseOverview(overview))
